physics is some sort of hell
