+++ src/js/ui.js
@@ -59,6 +59,184 @@ var ui = {
 	},
 	*/
 
+	joining: function(){
+		console.log("joining");
+
+    var data = ui.defaultPageData("search");
+    ui.handlebars("search", data);
+
+    ipcRenderer.send('find', 'local');
+		ui.menuchange('join');
+
+    ipcRenderer.on('found', function(event, services){
+      //ui.fadeSpinner();
+      console.log(services);
+      var data = ui.defaultPageData("search");
+      var hosts = [];
+
+      $.each(services, function( index, value ) {
+        var gameName = value.details.host_name+" Game";
+        var v = ui.contains(hosts, gameName, 'title' );
+        if(v < 0){
+          var details = value.details;
+          var host_details = {name: "host", value: JSON.stringify(details)};
+          hosts.push({id: index, class: "join-host", title: gameName, data: host_details})
+        }
+      });
+
+      $.each(hosts, function( index, value ) {
+        data.buttons.unshift(hosts[index]);
+      });
+
+      data.spinner = false;
+			ui.handlebars("search", data);
+    });
+
+		ipcRenderer.on('unfound', function(event, service){
+			console.log("unfound");
+      var data = ui.defaultPageData("search");
+      data.spinner = false;
+      //TODO: add message that nothing was found local or online?
+      ui.handlebars("search", data);
+		});
+	},
+
+	hosting: function(){
+		console.log("host");
+
+		var ip_address = ui.getAddress();
+
+		ipcRenderer.send('setup', ip_address);
+		console.log("setup");
+
+		ipcRenderer.on('hosting', function(event, service){
+			console.log("hosting");
+			//console.log(service);
+			ui.host = true;
+			service.host_name = nameUser;
+			ipcRenderer.send('advertise', service);
+			//console.log(service);
+			ui.common(service);
+
+		});
+
+	},
+
+	common: function(service){
+	  console.log("common started");
+
+	  io = require('socket.io-client'),
+	  socket = io.connect('http://'+service.ip+':'+service.port);
+
+	  socket.on('connect', function(){
+			socket.emit('newUser', uuid, nameUser);
+	  });
+
+		socket.on('createUser', function(data){
+			console.log("createUser");
+	    ui.user = data;
+	    console.log(data);
+	    users.push(data);
+	    ui.buildLobby(data);
+	    socket.emit('requestUsers', uuid);
+			if(pkg.debug){
+				ui.startMatch();
+			}else{
+				ui.menuchange('host');
+			}
+		});
+
+		socket.on('addUser', function(data){
+			console.log("addUser");
+	    //something is making this work weird
+			// var index = g.contains(users, data.playerId, 'playerId');
+			// if(index === -1){
+			// 	if(uuid != data.playerId){
+				users.push(data);
+				ui.addUser(data);
+			// 	}
+			// }
+
+		});
+
+	  socket.on('removeUser', function(id){
+	    console.log("removeUser");
+	    var users = [];
+	    var data = ui.user;
+	    ui.buildLobby(data);
+	    socket.emit('requestUsers', uuid);
+	  });
+
+	  socket.on('host-closed', function(data){
+	    console.log("host closed");
+	  });
+
+	  //game mode
+	  socket.on('startMatch', function(data, setup){
+	    if(!thisPlayer){
+	      if(data.playerId == uuid){
+
+	        ui.updateScore(setup);
+	        ui.showGUI(); // needs data to display
+	        g = new game;
+	        g.loadWorld(socket, data);
+
+	        socket.emit('requestPlayers', uuid);
+					socket.emit('requestCubes');
+
+	        ui.menuchange('game');
+	      }
+	    }
+	  });
+
+	  socket.on('addPlayer', function(data){
+	    g.addOtherPlayer(data);
+	  });
+
+		socket.on('addCubes', function(data){
+			for (var i = 0; i < data.length; i++){
+				g.addCube(data[i]);
+			}
+	    // var d = { x:0, y:0, z:0 };
+	    // g.addSmallCubes( d ); // testing
+		});
+
+	  socket.on('updatePlayers', function(data){
+	    //may need to check uuid and data.playerId dont match or data is not incorrect between
+	    g.updateObject(data);
+	  });
+
+		socket.on('updateShots', function(data){
+	    //console.log(data);
+			g.addShot(data);
+		});
+
+		socket.on('reportKill', function(data){
+	    console.log("reportKill");
+	    console.log(data);
+	    ui.updateScore(data);
+	    g.hit(data.kill);
+
+		});
+
+	  // socket.on('removePlayer', function(data){
+	  //   g.removeOtherPlayer(data);
+	  // });
+
+	  socket.on('endgame', function(data){
+	    console.log("endgame");
+	    console.log(data);
+	    //show winner message
+	    var promise = ui.fullText("Winner: "+data);
+	    promise.then(function(){
+	      //transition back to host
+	      g.deconstruct();
+	    });
+
+	  });
+
+	},
+
 	fullText: function(text, time){
 
 		var promise = new Promise(function( resolve ) {
@@ -150,6 +328,7 @@ var ui = {
 			ref.menuchange('main'); // may need to add a
 		});
 
+
 		$(document).on("click", "#stop-hosting", function(){
 			console.log("stop-hosting");
 
@@ -181,12 +360,16 @@ var ui = {
 		});
 
 		$(document).on("click", "#startMatch", function(){
-			console.log("startMatch");
-			//$( "#lobby" ).hide();updatePlayers
-			socket.emit('start');
+			ui.startMatch();
 		});
 
 		this.exitAppSetup();
+	},
+
+	startMatch: function (){
+		console.log("startMatch");
+		//$( "#lobby" ).hide();updatePlayers
+		socket.emit('start');
 	},
 
 	setl: function(term, data){
